+++ docs/docs/release-notes.md
@@ -2,6 +2,20 @@
 
 ## Latest Changes
 
+### 🚨 Breaking API Changes 
+
+** Renamed **
+
+*Functions & Methods:*
+
+* Renamed keyword argument `expected_checksum` of method `Calculator.verify` to `expected`
+
+### ✨ Added
+* Added support for other data types than `bytes` to `Calculator.checksum` and `Calculator.verify`
+ 
+     (`int`, `ByteString`, `BinaryIO`, `Iterable[ByteString]`)
+
+
 ## 2.0.0 - 2022-11-27
 
 ### 🚨 Breaking API Changes